+++ README.md
@@ -30,13 +30,13 @@
 
 ## 🕹 게임 방법
 
-![시작 페이지](https://s3-us-west-2.amazonaws.com/secure.notion-static.com/b0cc5db0-d54e-442e-9182-6ee18c6bfa9b/스크린샷_2022-03-11_오후_3.44.05.png)
+[시작 페이지]
 
 시작 페이지
 
 > 게임을 시작하기 전, How to play button을 통해 게임 조작 방법을 알 수 있습니다.
 
-![How to play modal - 게임 조작 방법](https://s3-us-west-2.amazonaws.com/secure.notion-static.com/17f37500-8e16-4225-b702-635aedb2b2f1/스크린샷_2022-03-11_오후_3.57.06.png)
+[How to play modal - 게임 조작 방법]
 
 How to play modal - 게임 조작 방법
 
@@ -49,13 +49,12 @@ How to play modal - 게임 조작 방법
 > 몬스터는 밟아 죽일 수 있지만,
 > 머리 외에 몸통이 먼저 닿게되면 게임 종료
 
-![Level select page - 난이도 선택 페이지](https://s3-us-west-2.amazonaws.com/secure.notion-static.com/2fa5305b-6702-4730-b872-437864ec33f3/스크린샷_2022-03-11_오후_3.51.44.png)
-
+[Level select page - 난이도 선택 페이지]
 Level select page - 난이도 선택 페이지
 
 > 게임의 난이도를 선택하여 진행할 수 있습니다.
 
-![Level 1 - ice map](https://s3-us-west-2.amazonaws.com/secure.notion-static.com/013407ed-ccd8-4098-a1aa-657eaa99b399/스크린샷_2022-03-11_오후_3.57.56.png)
+[Level 1 - ice map]
 
 Level 1 - ice map
 
@@ -66,14 +65,14 @@ Level 1 - ice map
 > Level이 높을 수록 몬스터가 많아지고,
 > 지형이 어려워집니다.
 
-![Level 2 - fire map](https://s3-us-west-2.amazonaws.com/secure.notion-static.com/b704bae9-ced1-476a-b7b2-99e2b9c4bfc8/스크린샷_2022-03-11_오후_4.00.29.png)
+[Level 2 - fire map]
 
 Level 2 - fire map
 
 > 우측 상단에는 Game의 bgm을
 > 끄고 킬 수 있는 on/off 버튼이 있습니다.
 
-![Level 3 - dark map](https://s3-us-west-2.amazonaws.com/secure.notion-static.com/df06c10d-5d9d-480b-8f0e-c3d67c6cc86a/스크린샷_2022-03-11_오후_3.58.26.png)
+[Level 3 - dark map]
 
 Level 3 - dark map
