+++ src/assets/images/monster/walkPurple.png
