+++ src/assets/images/obstacle/level1/iceLargeObstacle.png
