+++ src/assets/images/obstacle/level1/iceObstacle.png
