+++ src/assets/images/obstacle/level2/fireLargeObstacle.png
