+++ src/assets/images/obstacle/level2/fireObstacle.png
