+++ src/assets/images/obstacle/level3/darkLargeObstacle.png
