+++ src/assets/images/obstacle/level3/darkObstacle.png
