+++ src/assets/images/platform/level1/icePlatform.png
