+++ src/assets/images/platform/level1/iceSmallPlatform.png
