+++ src/assets/images/platform/level2/firePlatform.png
