+++ src/assets/images/platform/level3/darkPlatform.png
