+++ src/assets/images/platform/level3/darkSmallPlatform.png
