+++ src/assets/images/player/hurt.png
