+++ src/assets/images/player/idleLeft.png
