+++ src/assets/images/player/idleRight.png
