+++ src/assets/images/player/jumpingRight.png
