+++ src/index.css
@@ -1,5 +1,5 @@
 @font-face {
-	src: url("./assets//font/zig_____.ttf");
+	src: url("./assets/fonts/zig_____.ttf");
 	font-family: "arcade";
 }
 
@@ -259,7 +259,7 @@ body {
 	height: 100%;
 	justify-content: center;
 	align-items: center;
-	background-image: url("./assets/img/background/selecPage/selecPage.png");
+	background-image: url("./assets/images/background/selecPage/selecPage.png");
 	animation: movebg 5s linear infinite;
 }
