+++ src/index.html
@@ -27,11 +27,11 @@
 				<p class="modal-title">How To Play</p>
 				<div class="modal-content">
 					<div class="game-explain">
-						<img class="audio" src="./assets/img/audio/soft.png" alt="soft voice image" />Use
+						<img class="audio" src="./assets/images/audio/soft.png" alt="soft voice image" />Use
 						<span class="red-text">SOFT VOICE</span>to move
 						<img
 							class="arrow"
-							src="./assets/img/arrow/forward.png"
+							src="./assets/images/arrow/forward.png"
 							art="character motion explain"
 						/>
 					</div>
@@ -39,19 +39,23 @@
 					<div class="game-explain">
 						<img
 							class="audio-image"
-							src="./assets/img/audio/soft-loud.png"
+							src="./assets/images/audio/soft-loud.png"
 							alt="soft voice image"
 						/>Use <span class="red-text">LOUD VOICE</span>to jump
-						<img class="arrow" src="./assets/img/arrow/jump.png" art="character motion explain" />
+						<img
+							class="arrow"
+							src="./assets/images/arrow/jump.png"
+							art="character motion explain"
+						/>
 					</div>
 
 					<div class="game-explain">
-						<img class="audio-image" src="./assets/img/audio/loud.png" alt="soft voice image" />
+						<img class="audio-image" src="./assets/images/audio/loud.png" alt="soft voice image" />
 						Speck
 						<span class="red-text">LOUDER</span>, jump <span class="red-text">Higher</span>
 						<img
 							class="arrow-jump"
-							src="./assets/img/arrow/higher.png"
+							src="./assets/images/arrow/higher.png"
 							art="character motion explain"
 						/>
 					</div>
@@ -80,8 +84,8 @@
 		<div class="result-modal"></div>
 
 		<div class="toggle-music">
-			<img class="music-on" src="./assets/img/audio/on.png" alt="music-on" />
-			<img class="music-off" src="./assets/img/audio/off.png" alt="music-off" />
+			<img class="music-on" src="./assets/images/audio/on.png" alt="music-on" />
+			<img class="music-off" src="./assets/images/audio/off.png" alt="music-off" />
 		</div>
 
 		<div class="percent-container">
